% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_for_blank.R
\name{check_for_blank}
\alias{check_for_blank}
\title{\code{check_for_blank} Check to make sure there are no blank addresses in the address
column of the input file.}
\usage{
check_for_blank(addresses)
}
\arguments{
\item{addresses}{The dataframe column containing the addresses.}
}
\value{
Simply return the input addresses if there are no blanks.
}
\description{
\code{check_for_blank} Check to make sure there are no blank addresses in the address
column of the input file.
}
\details{
The API query will throw an error and stop the geocoding
process for the whole file if a URL is constructed with a blank address.This function
prevents wasted queries (only so many are allowed at no charge) by throwing an error
that tells the user to fix blank addresses.
}
\examples{
\dontrun{
check_for_blank(df['Business Address'])
}

}
