% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcode.R
\name{gcode}
\alias{gcode}
\title{\code{gcode} Geocodes a dataframe of addresses using Google Maps API via ggmap}
\usage{
gcode(df, components = FALSE)
}
\arguments{
\item{df}{A single-column dataframe of addresses.}

\item{components}{Logical. Postal code is included if \code{components = TRUE}. Defaults to FALSE.}
}
\value{
A dataframe object with initial address, lat, long and full address. Postal code may
also be included.
}
\description{
\code{gcode} Geocodes a dataframe of addresses using Google Maps API via ggmap
}
\details{
The function takes a single-column dataframe of addresses. It then
geocodes those addresses, returning a new dataframe with additional columns for latitude,
longitude and full address. Postal code can also be included as a separate column.
}
\examples{
\dontrun{
test_str <- head(test,3)

gcode(test_str, components = TRUE)
}
}
