% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcode_one.R
\name{gcode_one}
\alias{gcode_one}
\title{\code{gcode_one} Geocodes a single address using ggmap}
\usage{
gcode_one(address)
}
\arguments{
\item{address}{A single character object. The address to be geocoded.}
}
\value{
Returns a three element vector: c(lat,long, full_address)
}
\description{
\code{gcode_one} Geocodes a single address using ggmap
}
\details{
This function geocodes an address, returning a vector of information
containing latitude, longitude and full address.Google's API does not allow more
than 50 req/sec, so I force a pause of 1/40 of a sec just to be safe.
}
