% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_addresses.R
\name{get_addresses}
\alias{get_addresses}
\title{\code{get_addresses} Identifies the address column in a csv file, and returns it as a
data frame object.}
\usage{
get_addresses(infile, address_column = "address")
}
\arguments{
\item{infile}{A csv file to read as a character object. If in your cwd, you needn't specify
the filepath. If not, the file path should be attached.}

\item{address_column}{A character object. The name of the column containing the addresses.}
}
\value{
Returns the address column as a dataframe object.
}
\description{
\code{get_addresses} Identifies the address column in a csv file, and returns it as a
data frame object.
}
\examples{
\dontrun{
get_addresses('F:/gitrepos/R-sandbox/pei-test.csv','Business Address')

test <- get_addresses("pei_test.csv", address_column = "Business Address")
}
}
