% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_components.R
\name{get_components}
\alias{get_components}
\title{\code{get_components} A function that appends separate components, like postal code and province, as new variables
to a geocoded df.}
\usage{
get_components(df)
}
\arguments{
\item{df}{A dataframe object with at least one column that contains the full addresses from which
components are extracted.}
}
\value{
A dataframe with full address and components (postal code, prov,..) and potentially other
data.
}
\description{
\code{get_components} A function that appends separate components, like postal code and province, as new variables
to a geocoded df.
}
\details{
The function extracts the postal code (and later the province) from the full address
column of the input dataframe. It appends these components to the df as separate variables, and
returns a new dataset
}
\examples{
\dontrun{
get_components(coded)
}
}
