% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_key.R
\name{register_key}
\alias{register_key}
\title{\code{register_key} Register's your API key needed to geocode. If the key is not registered,
The geocoding functions will throw errors.}
\usage{
register_key(api_key)
}
\arguments{
\item{api_key}{The api key, as a character object.}
}
\description{
\code{register_key} Register's your API key needed to geocode. If the key is not registered,
The geocoding functions will throw errors.
}
\examples{
\dontrun{
register_key('ASnERN405060i7NDNS') # This key is fictional
}
}
