% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_weird_characters.R
\name{remove_weird_characters}
\alias{remove_weird_characters}
\title{\code{remove_weird_characters} Removes certain characters from entries in the address column.}
\usage{
remove_weird_characters(addresses)
}
\arguments{
\item{addresses}{The dataframe column containing the addresses.}
}
\value{
returns the input dataframe column with special characters replaced by whitespace
}
\description{
\code{remove_weird_characters} Removes certain characters from entries in the address column.
}
\details{
Some characters, like '#', '', or '.' can throw off Google's API when included in the
http request, since they serve specific purposes in dividing the URL. To avoid error, this
function removes those special characters.
}
\examples{
\dontrun{
remove_weird_characters(as.data.frame("720 Bathurst St #411, Toronto, ON M5S 2R4"))
remove_weird_characters(data['address'])
}
}
