% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_file.R
\name{write_to_file}
\alias{write_to_file}
\title{\code{write_to_file} - Writes a dataframe as a CSV file.}
\usage{
write_to_file(df, name)
}
\arguments{
\item{df}{The dataframe to write.}

\item{name}{The name of the output file, as a character object. If a file path is not given,
The function assumes cwd.}
}
\description{
\code{write_to_file} - Writes a dataframe as a CSV file.
}
